Feat : 같이 게임하기 기능 추가
