+++ src/util/util.ts
@@ -77,3 +77,23 @@ export const titleAction = async (navigate: NavigateFunction, id: string) => {
     }
   }
 };
+
+// 랜덤 방제목 정하는 함수
+export const randomNameFunc = () => {
+  const data = [
+    '캐치마인드 신나는 한 판~!',
+    '아이브 영원하라',
+    '그림 고수분만',
+    '초보만 들어오세요',
+    '중랑 12여 남자만',
+    '15살 수다방',
+    '김가을양재혁',
+    '안유진은 양재혁 여친',
+    '살려주세요',
+    '패캠 뒷담방 ex) 돈 내놔....',
+  ];
+
+  const randomIndex = Math.floor(Math.random() * data.length);
+  const randomPick = data[randomIndex];
+  return randomPick;
+};
